requirements.txt
python-telegram-bot==13.15
python-dotenv==0.21.0
requests==2.31.0
